backend/requirements.txt
aiofiles==23.1.0
aiotinydb==2.0.0
anyio==3.6.2
appdirs==1.4.4
async-lru==1.0.3
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cssselect==1.2.0
fake-useragent==1.1.1
fastapi==0.92.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
LivePopularTimes==1.3
lxml==4.9.2
parse==1.19.0
Pillow==9.4.0
pydantic==1.10.5
pyee==8.2.2
PyJWT==2.6.0
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
requests-html==0.10.0
requests==2.28.2
rfc3986==1.5.0
sentry-sdk==1.15.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.25.0
tinydb==4.7.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
w3lib==2.1.1
watchfiles==0.18.1
websockets==10.4
zipp==3.13.0
